refactor: type refactoring + ts-ignore rsolved | mz
