+++ src/styles/types.ts
@@ -1,10 +1,15 @@
-import type { Theme } from '.';
+import type { Theme } from '../styles';
 import destinationNetworks from '../data/destinationNetworks.json';
+import { CONTRACT_ADDRESSES } from '../helpers';
 
 export type ThemeProps = {
 	theme: Theme;
 };
 
-export type networks = typeof destinationNetworks;
+export type ColorType = 'default' | 'error' | 'warning' | 'icon' | 'success';
+
+export type Networks = typeof destinationNetworks;
+export type DestinationNetworks = keyof Networks;
+export type ContractAdress = keyof typeof CONTRACT_ADDRESSES;
 
 export type ApiAuthType = { access: string; is_kyced: boolean; refresh: string };
